# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PetCare",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://petcare-production-cc6b.up.railway.app/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        },
        {
            "url": "https://petcare-production-3202.up.railway.app/api"
        }
    ],
    "paths": {
...
        "/contact-messages": {
            "post": {
                "tags": [
                    "User - Contact Us"
                ],
                "description": "Sent new message.",
                "operationId": "add_message",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}